add some memory pool structs

+++ src-tauri/src/damselfly/memory/memory_cache.rs
@@ -83,6 +83,6 @@ mod tests {
         let memory_updates = mst_parser.parse_log("./trace4.log", "./threadxApp").memory_updates;
         let update_interval_factory = UpdateIntervalFactory::new(memory_updates);
         let update_intervals = update_interval_factory.construct_enum_vector();
-        let memory_cache = MemoryCache::new(4, update_intervals, DEFAULT_CACHE_INTERVAL);
+        let memory_cache = MemoryCache::new(4, update_intervals, DEFAULT_CACHE_INTERVAL as usize);
     }
 }
